Move name mapper and related into separate package
