+++ src/main/scala/net/ceedubs/ficus/readers/namemappers/DefaultNameMapper.scala
@@ -1,0 +1,10 @@
+package net.ceedubs.ficus.readers.namemappers
+
+/**
+  * Default implementation for name mapper, names in code equivalent to names in configuration
+  */
+case object DefaultNameMapper extends NameMapper {
+
+  override def map(name: String): String = name
+
+}
